Refactors model to it's own file
